fix + refactor: namespaced plugin files and made package path edit cleaner
